refactor: 디렉토리 구조 변경에 따른 경로 수정
